+++ packages/appinio_video_player/CHANGELOG.md
@@ -1,7 +1,12 @@
+## [1.1.0] - 2022.04.06
+
+- Removed dependencies from other packages (provider, wakelock)
+- Restructuring
+
 ## [1.0.1] - 2021.10.05
 
-* Bugfix: dispose video values
+- Bugfix: dispose video values
 
 ## [1.0.0] - 2021.10.04
 
-* Initial version 
+- Initial version